+++ src/main/java/pl/grizzlysoftware/dotykacka/client/v1/api/service/EmployeeService.java
@@ -11,23 +11,23 @@ import java.util.Collection;
  */
 public interface EmployeeService {
     @GET("{cloudId}/{id}")
-    Call<Employee> getEmployee(@Path("cloudId") Integer cloudId, @Path("id") Long id);
+    Call<Employee> getEmployee(@Path("cloudId") Long cloudId, @Path("id") Long id);
 
     @GET("{cloudId}")
-    Call<Collection<Employee>> getEmployees(@Path("cloudId") Integer cloudId, @Query("limit") Integer limit, @Query("offset") Integer offset, @Query("sort") String sort);
+    Call<Collection<Employee>> getEmployees(@Path("cloudId") Long cloudId, @Query("limit") Integer limit, @Query("offset") Integer offset, @Query("sort") String sort);
 
     @POST("{cloudId}/{id}/update")
     @Headers({
             "Content-Type: application/json",
             "Accept: application/json"})
-    Call<Employee> updateEmployee(@Path("cloudId") Integer cloudId, @Path("id") Long id, @Body Employee employee);
+    Call<Employee> updateEmployee(@Path("cloudId") Long cloudId, @Path("id") Long id, @Body Employee employee);
 
     @GET("{cloudId}/{id}/delete")
-    Call<Employee> deleteEmployee(@Path("cloudId") Integer cloudId, @Path("id") Long id, @Query("anonymize") Boolean anonymizeEmployee);
+    Call<Employee> deleteEmployee(@Path("cloudId") Long cloudId, @Path("id") Long id, @Query("anonymize") Boolean anonymizeEmployee);
 
     @Headers({
             "Content-Type: application/json",
             "Accept: application/json"})
     @POST("{cloudId}/create")
-    Call<Employee> createEmployee(@Path("cloudId") Integer cloudId, @Body Employee employee);
+    Call<Employee> createEmployee(@Path("cloudId") Long cloudId, @Body Employee employee);
 }
